requirements.txt
line-bot-sdk
Flask
mecab-python3
keras
tensorflow
opencv-python
numpy
pandas
